requirements.txt
langchain
openai
streamlit
tiktoken
unstructured
pdf2image
pdfminer
PyPDF2
python-dotenv
ipykernel
faiss-cpu
huggingface_hub
pinecone-client
cassio
pypdf
pandas
numpy
matplotlib